Simplify BufferInfo
Flatten and remove duplicated information
