+++ core/trino-main/src/test/java/io/trino/execution/buffer/TestArbitraryOutputBuffer.java
@@ -1030,15 +1030,13 @@ public class TestArbitraryOutputBuffer
                 bufferInfo,
                 new BufferInfo(
                         bufferId,
-                        false,
+                        // every page has one row
+                        bufferedPages + pagesSent,
+                        bufferedPages + pagesSent,
                         bufferedPages,
+                        sizeOfPages(bufferedPages).toBytes(),
                         pagesSent,
-                        new PageBufferInfo(
-                                bufferId.getId(),
-                                bufferedPages,
-                                sizeOfPages(bufferedPages).toBytes(),
-                                bufferedPages + pagesSent, // every page has one row
-                                bufferedPages + pagesSent)));
+                        false));
     }
 
     @SuppressWarnings("ConstantConditions")
